Fix one edge case in the proxy.
This change updates the logic in the proxy that
forwards bytes between sockets to fix one edge
case where timeouts were not being handled
correctly.

More importantly, this change also adds a unit test
that exercises the proxy logic with various
combinations of timeouts.

Finally, this change also updates the requirements
which were out of date and prevented the tests
from running successfully in a clean environment.